* Fix NPC enemy selection after respawn. * Clean up some NPC code. * New automated system to switch between spin projectile blocking mode and saber blocking mode.

+++ codemp/win32/win_main.cpp
@@ -22,7 +22,7 @@ char *Sys_BinaryPath(void);
 
 extern "C"
 {
-	__declspec(dllexport) unsigned long NvOptimusEnablement = 0x00000001;
+	__declspec(dllexport) DWORD NvOptimusEnablement = 0x00000001;
 	__declspec(dllexport) int AmdPowerXpressRequestHighPerformance = 1;
 }
